Added custom class for sensor
