+++ components/maidsite_desk/number/__init__.py
@@ -41,10 +41,13 @@ CONFIG_SCHEMA = cv.Schema(
 async def to_code(config):
   hub = await cg.get_variable(config[CONF_MAIDSITE_DESK_ID])
   for index, key in enumerate(TYPES):
-    await setup_conf(config, key, index, hub)
-
-async def setup_conf(parent_config, key, index, hub):
-  if child_config := parent_config.get(key):
-    var = await number.new_number(child_config, min_value=CONF_NUMBERS[index][0], max_value=CONF_NUMBERS[index][1], step=CONF_NUMBERS[index][2])
-    await cg.register_parented(var, parent_config[CONF_MAIDSITE_DESK_ID])
-    cg.add(getattr(hub, f"set_{key}_number")(var))
+    if child_config := config.get(key):
+      var = await number.new_number(
+        child_config,
+        min_value=CONF_NUMBERS[index][0],
+        max_value=CONF_NUMBERS[index][1],
+        step=CONF_NUMBERS[index][2]
+      )
+      await cg.register_component(var, child_config)
+      await cg.register_parented(var, config[CONF_MAIDSITE_DESK_ID])
+      cg.add(getattr(hub, f"set_{key}_number")(var))
