+++ components/maidsite_desk/number/maidsite_desk_number.cpp
@@ -1,10 +1,18 @@
 #include "maidsite_desk_number.h"
+#include "esphome/core/log.h"
 
 
 namespace esphome
 {
   namespace maidsite_desk
   {
+    static const char *const TAG = "maidsite_desk.number";
+
+    void MaidsiteDeskNumber::dump_config()
+    {
+      LOG_NUMBER("", "Maidsite Desk Number", this);
+    }
+
     void MaidsiteDeskNumber::control(float value)
     {
       this->publish_state(value);
