+++ components/maidsite_desk/number/maidsite_desk_number.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "esphome/components/number/number.h"
+#include "esphome/core/component.h"
 #include "../maidsite_desk.h"
 
 
@@ -7,10 +8,12 @@ namespace esphome
 {
   namespace maidsite_desk
   {
+    class MaidsiteDeskComponent;
     class MaidsiteDeskNumber : public number::Number, public Parented<MaidsiteDeskComponent>
     {
     public:
       MaidsiteDeskNumber() = default;
+      void dump_config() override;
 
     protected:
       void control(float value) override;
