+++ components/maidsite_desk/sensor/maidsite_desk_sensor.cpp
@@ -1,0 +1,16 @@
+#include "maidsite_desk_sensor.h"
+#include "esphome/core/log.h"
+
+
+namespace esphome
+{
+  namespace maidsite_desk
+  {
+    static const char *const TAG = "maidsite_desk.sensor";
+
+    void MaidsiteDeskSensor::dump_config()
+    {
+      LOG_SENSOR("", "Maidsite Desk Sensor", this);
+    }
+  } // namespace maidsite_desk
+} // namespace esphome
